Move geometry functionality to separate directory
