+++ include/pbrt/core/geometry/Ray.hpp
@@ -1,8 +1,8 @@
 #pragma once
 
-#include "core.hpp"
-#include "math/Point3.hpp"
-#include "math/Vector3.hpp"
+#include "pbrt/core/core.hpp"
+#include "pbrt/core/math/Point3.hpp"
+#include "pbrt/core/math/Vector3.hpp"
 
 namespace idragnev::pbrt {
     class Ray
