+++ src/core/Shape.cpp
@@ -1,6 +1,6 @@
 #include "pbrt/core/Shape.hpp"
 #include "pbrt/core/transformations/Transformation.hpp"
-#include "pbrt/core/Bounds3.hpp"
+#include "pbrt/core/geometry/Bounds3.hpp"
 
 namespace idragnev::pbrt {
     Shape::Shape(const Transformation& objectToWorld,
