+++ src/core/TransformedPrimitive.cpp
@@ -1,7 +1,7 @@
 #include "pbrt/core/TransformedPrimitive.hpp"
 #include "pbrt/core/transformations/Transformation.hpp"
 #include "pbrt/core/SurfaceInteraction.hpp"
-#include "pbrt/core/Bounds3.hpp"
+#include "pbrt/core/geometry/Bounds3.hpp"
 
 #include <assert.h>
