+++ src/core/transformations/Transformation.cpp
@@ -1,7 +1,7 @@
 #include "pbrt/core/transformations/Transformation.hpp"
 #include "pbrt/core/math/Vector3.hpp"
 #include "pbrt/core/math/Point3.hpp"
-#include "pbrt/core/Bounds3.hpp"
+#include "pbrt/core/geometry/Bounds3.hpp"
 #include "pbrt/core/SurfaceInteraction.hpp"
 
 namespace idragnev::pbrt {
