+++ src/shapes/Disk.cpp
@@ -1,7 +1,7 @@
 #include "pbrt/shapes/Disk.hpp"
 #include "pbrt/core/transformations/Transformation.hpp"
 #include "pbrt/core/EFloat.hpp"
-#include "pbrt/core/Bounds3.hpp"
+#include "pbrt/core/geometry/Bounds3.hpp"
 #include "pbrt/core/math/Math.hpp"
 
 namespace idragnev::pbrt::shapes {
