+++ src/shapes/Paraboloid.cpp
@@ -1,5 +1,5 @@
 #include "pbrt/shapes/Paraboloid.hpp"
-#include "pbrt/core/Bounds3.hpp"
+#include "pbrt/core/geometry/Bounds3.hpp"
 #include "pbrt/core/EFloat.hpp"
 #include "pbrt/core/transformations/Transformation.hpp"
