+++ src/shapes/Triangle.cpp
@@ -7,7 +7,7 @@
 #include "pbrt/core/math/Normal3.hpp"
 #include "pbrt/core/Texture.hpp"
 #include "pbrt/functional/Functional.hpp"
-#include "pbrt/core/Bounds3.hpp"
+#include "pbrt/core/geometry/Bounds3.hpp"
 
 namespace idragnev::pbrt::shapes {
     TriangleMesh::TriangleMesh(
